+++ example/lib/providers/payment_provider.dart
@@ -4,10 +4,7 @@ import 'package:amazon_payfort_example/constants/fort_constants.dart';
 import 'package:amazon_payfort_example/models/sdk_token_response.dart';
 import 'package:amazon_payfort_example/providers/default_change_notifier.dart';
 import 'package:network_info_plus/network_info_plus.dart';
-
-typedef PaymentCompleteCallback = void Function(PayFortResult result);
-
-typedef PaymentErrorCallback = void Function(String message);
+import 'package:uuid/uuid.dart';
 
 class PaymentProvider extends DefaultChangeNotifier {
   final AmazonPayfort _payfort = AmazonPayfort.instance;
@@ -17,13 +14,14 @@ class PaymentProvider extends DefaultChangeNotifier {
   Future<void> init() async {
     /// Step 1:  Initialize Amazon Payfort
     await AmazonPayfort.initialize(
-      PayFortOptions(environment: FortConstants.environment),
+      const PayFortOptions(environment: FortConstants.environment),
     );
   }
 
-  Future<void> paymentWithCraditOrDebitCard({
-    required PaymentCompleteCallback onCompleted,
-    required PaymentErrorCallback onError,
+  Future<void> paymentWithCreditOrDebitCard({
+    required SucceededCallback onSucceeded,
+    required FailedCallback onFailed,
+    required CancelledCallback onCancelled,
   }) async {
     try {
       var sdkTokenResponse = await _generateSdkToken();
@@ -35,44 +33,54 @@ class PaymentProvider extends DefaultChangeNotifier {
         customerEmail: 'test@customer.com',
         orderDescription: 'Test Order',
         sdkToken: sdkTokenResponse?.sdkToken ?? '',
-        merchantReference: 'Order ${DateTime.now().millisecondsSinceEpoch}',
+        merchantReference: const Uuid().v4(),
         currency: 'SAR',
-        customerIp: await _info.getWifiIP(),
+        customerIp: (await _info.getWifiIP() ?? ''),
       );
 
-      var payfortResult = await _payfort.callPayFort(request);
-      onCompleted(payfortResult);
+      _payfort.callPayFort(
+        request: request,
+        callBack: PayFortResultCallback(
+          onSucceeded: onSucceeded,
+          onFailed: onFailed,
+          onCancelled: onCancelled,
+        ),
+      );
     } catch (e) {
-      onError(e.toString());
+      onFailed(e.toString());
     }
   }
 
   Future<void> paymentWithApplePay({
-    required PaymentCompleteCallback onCompleted,
-    required PaymentErrorCallback onError,
+    required SucceededCallback onSucceeded,
+    required FailedCallback onFailed,
   }) async {
     try {
       var sdkTokenResponse = await _generateSdkToken(isApplePay: true);
 
-      /// Step 4: Processing Payment
+      /// Step 4: Processing Payment [Don't multiply with 100]
       FortRequest request = FortRequest(
-        amount: 10,
+        amount: 1000,
         customerName: 'Test Customer',
         customerEmail: 'test@customer.com',
         orderDescription: 'Test Order',
         sdkToken: sdkTokenResponse?.sdkToken ?? '',
-        merchantReference: 'Order ${DateTime.now().millisecondsSinceEpoch}',
+        merchantReference: const Uuid().v4(),
         currency: 'SAR',
-        customerIp: await _info.getWifiIP(),
+        customerIp: (await _info.getWifiIP() ?? ''),
       );
 
-      var payfortResult = await _payfort.callPayFortForApplePay(
+      _payfort.callPayFortForApplePay(
         request: request,
+        countryIsoCode: 'SA',
         applePayMerchantId: FortConstants.applePayMerchantId,
+        callback: ApplePayResultCallback(
+          onSucceeded: onSucceeded,
+          onFailed: onFailed,
+        ),
       );
-      onCompleted(payfortResult);
     } catch (e) {
-      onError(e.toString());
+      onFailed(e.toString());
     }
   }