+++ ios/Classes/SwiftAmazonPayfortPlugin.swift
@@ -1,11 +1,11 @@
 import Flutter
 import UIKit
 
-public class SwiftAmazonPayfortPlugin: NSObject, FlutterPlugin, ApplePayResponseDelegateProtocol {
+public class SwiftAmazonPayfortPlugin: NSObject, FlutterPlugin {
     
-    private var result :FlutterResult? = nil
+    public var delegate = PayFortDelegate()
     
-    private var fortDelegate = PayFortDelegate()
+    private static var channel:  FlutterMethodChannel? = nil
     
     public static func register(with registrar: FlutterPluginRegistrar) {
         
@@ -14,25 +14,25 @@ public class SwiftAmazonPayfortPlugin: NSObject, FlutterPlugin, ApplePayResponse
         let instance = SwiftAmazonPayfortPlugin()
         
         registrar.addMethodCallDelegate(instance, channel: channel)
+        
+        self.channel = channel
+        
     }
     
-    
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-        self.result = result
         
         if call.method == "initialize" {
             
             if let arguments = call.arguments as? Dictionary<String, Any>{
                 let options = processPayFortOptions(arguments: arguments)
-                fortDelegate.initialize(options: options)
-                fortDelegate.responseDelegate = self
+                delegate.initialize(options: options, channel: SwiftAmazonPayfortPlugin.channel!)
                 result(true)
             }
             result(false)
             
         } else if call.method == "getUDID" {
             
-            let udid = fortDelegate.getUDID()
+            let udid = delegate.getUDID()
             result(udid)
             
         } else if call.method == "generateSignature" {
@@ -40,7 +40,7 @@ public class SwiftAmazonPayfortPlugin: NSObject, FlutterPlugin, ApplePayResponse
             if let args = call.arguments as? Dictionary<String, Any>{
                 _ = args["shaType"] as? String
                 let string = args["concatenatedString"] as? String
-                let signature = fortDelegate.generateSignature(concatenatedString: string)
+                let signature = delegate.generateSignature(concatenatedString: string)
                 result(signature)
             }
             
@@ -48,17 +48,14 @@ public class SwiftAmazonPayfortPlugin: NSObject, FlutterPlugin, ApplePayResponse
             
             if let requestData = call.arguments as? Dictionary<String, Any>{
                 let viewController = UIApplication.shared.keyWindow?.rootViewController ?? UIViewController()
-                fortDelegate.callPayFort(requestData: requestData, viewController: viewController) { response in
-                    result(response)
-                }
-                
+                delegate.callPayFort(requestData: requestData, viewController: viewController)
             }
             
         } else if call.method == "callPayFortForApplePay" {
             
             if let requestData = call.arguments as? Dictionary<String, Any>{
                 let viewController = UIApplication.shared.keyWindow?.rootViewController ?? UIViewController()
-                fortDelegate.callPayFortForApplePay(requestData: requestData, viewController: viewController)
+                delegate.callPayFortForApplePay(requestData: requestData, viewController: viewController)
             }
             
         } else {
@@ -66,20 +63,17 @@ public class SwiftAmazonPayfortPlugin: NSObject, FlutterPlugin, ApplePayResponse
         }
     }
     
-    func onApplePayPaymentResponse(response: [String : Any]) {
-        result?(response)
-    }
-    
     
     private func processPayFortOptions(arguments: Dictionary<String, Any>) -> PayFortOptions {
         let options =  PayFortOptions(
-            environment : arguments["environment"] as? String ?? "default value",
+            environment : arguments["environment"] as? String ?? "",
             hideLoading: arguments["hideLoading"] as? Bool ?? false,
             presentAsDefault: arguments["presentAsDefault"] as? Bool ?? true,
             isShowResponsePage :arguments["isShowResponsePage"] as? Bool ?? true
         )
         return options
     }
+    
 }